Merges date casting tools into cast file
