+++ transform/cast.go
@@ -20,6 +20,11 @@ import (
 	"time"
 )
 
+const (
+	dateInputFormat  = "20060102"
+	dateOutputFormat = "2006-01-02"
+)
+
 func toInt(v string) (*int, error) {
 	if v == "" {
 		return nil, nil
@@ -59,6 +64,27 @@ func toBool(v string) *bool {
 		return nil
 	}
 	return &b
+}
+
+type date time.Time
+
+func (d *date) UnmarshalJSON(b []byte) error {
+	s := strings.Trim(string(b), `"`)
+	if s == "" {
+		return nil
+	}
+	t, err := time.Parse(dateOutputFormat, s)
+	if err != nil {
+		return err
+	}
+
+	*d = date(t)
+	return nil
+}
+
+func (d *date) MarshalJSON() ([]byte, error) {
+	t := time.Time(*d)
+	return []byte(`"` + t.Format(dateOutputFormat) + `"`), nil
 }
 
 // toDate expects a date as string in the format YYYYMMDD (that is the format
